+++ src/rnn/rnn.h
@@ -161,7 +161,7 @@ public:
 
     Expr output;
     Expr layerInput = input;
-    for(int i = 0; i < rnns_.size(); ++i) {
+    for(size_t i = 0; i < rnns_.size(); ++i) {
       auto lazyInput = layerInput;
 
       auto cell = rnns_[i]->at(0);
@@ -188,7 +188,7 @@ public:
 
     Expr output;
     Expr layerInput = input;
-    for(int i = 0; i < rnns_.size(); ++i) {
+    for(size_t i = 0; i < rnns_.size(); ++i) {
       Expr lazyInput;
       auto cell = rnns_[i]->at(0);
       auto lazyInputs = cell->getLazyInputs(shared_from_this());
@@ -217,7 +217,7 @@ public:
 
     Expr output;
     Expr layerInput = input;
-    for(int i = 0; i < rnns_.size(); ++i) {
+    for(size_t i = 0; i < rnns_.size(); ++i) {
       auto lazyInput = layerInput;
 
       auto cell = rnns_[i]->at(0);
